move tests to instances
