+++ lib/crypto/src/poseidon2/instance/vesta.rs
@@ -292,3 +292,55 @@ lazy_static! {
     pub static ref POSEIDON2_VESTA_PARAMS: Arc<Poseidon2Params<Scalar>> =
         Arc::new(Poseidon2Params::new(3, 5, 8, 56, &MAT_DIAG3_M_1, &MAT_INTERNAL3, &RC3));
 }
+
+#[allow(unused_imports)]
+#[cfg(test)]
+mod poseidon2_tests_vesta {
+    use crate::{
+        field::instance::FpVesta,
+        fp_from_hex,
+        poseidon2::{instance::vesta::POSEIDON2_VESTA_PARAMS, *},
+    };
+
+    type Scalar = FpVesta;
+
+    static TESTRUNS: usize = 5;
+
+    #[test]
+    fn consistent_perm() {
+        let poseidon2 = Poseidon2::new(&POSEIDON2_VESTA_PARAMS);
+        let t = poseidon2.params.t;
+        for _ in 0..TESTRUNS {
+            let input1: Vec<Scalar> = (0..t).map(|_| random_scalar()).collect();
+
+            let mut input2: Vec<Scalar>;
+            loop {
+                input2 = (0..t).map(|_| random_scalar()).collect();
+                if input1 != input2 {
+                    break;
+                }
+            }
+
+            let perm1 = poseidon2.permutation(&input1);
+            let perm2 = poseidon2.permutation(&input1);
+            let perm3 = poseidon2.permutation(&input2);
+            assert_eq!(perm1, perm2);
+            assert_ne!(perm1, perm3);
+        }
+    }
+
+    #[test]
+    fn kats() {
+        let poseidon2 = Poseidon2::new(&POSEIDON2_VESTA_PARAMS);
+        let mut input: Vec<Scalar> = vec![];
+        for i in 0..poseidon2.params.t {
+            input.push(Scalar::from(i as u64));
+        }
+        let perm = poseidon2.permutation(&input);
+        assert_eq!(perm[0], fp_from_hex!("261ecbdfd62c617b82d297705f18c788fc9831b14a6a2b8f61229bef68ce2792"));
+        assert_eq!(perm[1], fp_from_hex!("2c76327e0b7653873263158cf8545c282364b183880fcdea93ca8526d518c66f"));
+        assert_eq!(perm[2], fp_from_hex!("262316c0ce5244838c75873299b59d763ae0849d2dd31bdc95caf7db1c2901bf"));
+    }
+
+    // TODO#q: test against `zkhash` implementation should be added
+}
