+++ lib/crypto/src/poseidon2/mod.rs
@@ -245,483 +245,6 @@ impl<F: PrimeField> MerkleTreeHash<F> for Poseidon2<F> {
 
 #[allow(unused_imports)]
 #[cfg(test)]
-mod poseidon2_tests_goldilocks {
-    use super::*;
-    use crate::{
-        field::instance::FpGoldiLocks,
-        fp_from_hex,
-        poseidon2::instance::goldilocks::{
-            POSEIDON2_GOLDILOCKS_12_PARAMS, POSEIDON2_GOLDILOCKS_16_PARAMS,
-            POSEIDON2_GOLDILOCKS_20_PARAMS, POSEIDON2_GOLDILOCKS_8_PARAMS,
-        },
-    };
-
-    type Scalar = FpGoldiLocks;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let instances = vec![
-            Poseidon2::new(&POSEIDON2_GOLDILOCKS_8_PARAMS),
-            Poseidon2::new(&POSEIDON2_GOLDILOCKS_12_PARAMS),
-            Poseidon2::new(&POSEIDON2_GOLDILOCKS_16_PARAMS),
-            Poseidon2::new(&POSEIDON2_GOLDILOCKS_20_PARAMS),
-        ];
-        for instance in instances {
-            let t = instance.params.t;
-            for _ in 0..TESTRUNS {
-                let input1: Vec<Scalar> =
-                    (0..t).map(|_| random_scalar()).collect();
-
-                let mut input2: Vec<Scalar>;
-                loop {
-                    input2 = (0..t).map(|_| random_scalar()).collect();
-                    if input1 != input2 {
-                        break;
-                    }
-                }
-
-                let perm1 = instance.permutation(&input1);
-                let perm2 = instance.permutation(&input1);
-                let perm3 = instance.permutation(&input2);
-                assert_eq!(perm1, perm2);
-                assert_ne!(perm1, perm3);
-            }
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_GOLDILOCKS_12_PARAMS);
-        let mut input: Vec<Scalar> = vec![];
-        for i in 0..poseidon2.params.t {
-            input.push(Scalar::from(i as u64));
-        }
-        let perm = poseidon2.permutation(&input);
-        assert_eq!(perm[0], fp_from_hex!("01eaef96bdf1c0c1"));
-        assert_eq!(perm[1], fp_from_hex!("1f0d2cc525b2540c"));
-        assert_eq!(perm[2], fp_from_hex!("6282c1dfe1e0358d"));
-        assert_eq!(perm[3], fp_from_hex!("e780d721f698e1e6"));
-        assert_eq!(perm[4], fp_from_hex!("280c0b6f753d833b"));
-        assert_eq!(perm[5], fp_from_hex!("1b942dd5023156ab"));
-        assert_eq!(perm[6], fp_from_hex!("43f0df3fcccb8398"));
-        assert_eq!(perm[7], fp_from_hex!("e8e8190585489025"));
-        assert_eq!(perm[8], fp_from_hex!("56bdbf72f77ada22"));
-        assert_eq!(perm[9], fp_from_hex!("7911c32bf9dcd705"));
-        assert_eq!(perm[10], fp_from_hex!("ec467926508fbe67"));
-        assert_eq!(perm[11], fp_from_hex!("6a50450ddf85a6ed"));
-    }
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
-mod poseidon2_tests_babybear {
-    use std::convert::TryFrom;
-
-    use super::*;
-    use crate::{
-        field::instance::FpBabyBear,
-        fp_from_hex,
-        poseidon2::instance::babybear::{
-            POSEIDON2_BABYBEAR_16_PARAMS, POSEIDON2_BABYBEAR_24_PARAMS,
-        },
-    };
-
-    type Scalar = FpBabyBear;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let instances = vec![
-            Poseidon2::new(&POSEIDON2_BABYBEAR_16_PARAMS),
-            Poseidon2::new(&POSEIDON2_BABYBEAR_24_PARAMS),
-        ];
-        for instance in instances {
-            let t = instance.params.t;
-            for _ in 0..TESTRUNS {
-                let input1: Vec<Scalar> =
-                    (0..t).map(|_| random_scalar()).collect();
-
-                let mut input2: Vec<Scalar>;
-                loop {
-                    input2 = (0..t).map(|_| random_scalar()).collect();
-                    if input1 != input2 {
-                        break;
-                    }
-                }
-
-                let perm1 = instance.permutation(&input1);
-                let perm2 = instance.permutation(&input1);
-                let perm3 = instance.permutation(&input2);
-                assert_eq!(perm1, perm2);
-                assert_ne!(perm1, perm3);
-            }
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_BABYBEAR_24_PARAMS);
-        let mut input: Vec<Scalar> = vec![];
-        for i in 0..poseidon2.params.t {
-            input.push(Scalar::from(i as u64));
-        }
-        let perm = poseidon2.permutation(&input);
-        assert_eq!(perm[0], fp_from_hex!("2ed3e23d"));
-        assert_eq!(perm[1], fp_from_hex!("12921fb0"));
-        assert_eq!(perm[2], fp_from_hex!("0e659e79"));
-        assert_eq!(perm[3], fp_from_hex!("61d81dc9"));
-        assert_eq!(perm[4], fp_from_hex!("32bae33b"));
-        assert_eq!(perm[5], fp_from_hex!("62486ae3"));
-        assert_eq!(perm[6], fp_from_hex!("1e681b60"));
-        assert_eq!(perm[7], fp_from_hex!("24b91325"));
-        assert_eq!(perm[8], fp_from_hex!("2a2ef5b9"));
-        assert_eq!(perm[9], fp_from_hex!("50e8593e"));
-        assert_eq!(perm[10], fp_from_hex!("5bc818ec"));
-        assert_eq!(perm[11], fp_from_hex!("10691997"));
-        assert_eq!(perm[12], fp_from_hex!("35a14520"));
-        assert_eq!(perm[13], fp_from_hex!("2ba6a3c5"));
-        assert_eq!(perm[14], fp_from_hex!("279d47ec"));
-        assert_eq!(perm[15], fp_from_hex!("55014e81"));
-        assert_eq!(perm[16], fp_from_hex!("5953a67f"));
-        assert_eq!(perm[17], fp_from_hex!("2f403111"));
-        assert_eq!(perm[18], fp_from_hex!("6b8828ff"));
-        assert_eq!(perm[19], fp_from_hex!("1801301f"));
-        assert_eq!(perm[20], fp_from_hex!("2749207a"));
-        assert_eq!(perm[21], fp_from_hex!("3dc9cf21"));
-        assert_eq!(perm[22], fp_from_hex!("3c985ba2"));
-        assert_eq!(perm[23], fp_from_hex!("57a99864"));
-    }
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
-mod poseidon2_tests_bls12 {
-    use std::convert::TryFrom;
-
-    use super::*;
-    use crate::{
-        field::instance::FpBLS12,
-        fp_from_hex,
-        poseidon2::instance::bls12::{
-            POSEIDON2_BLS_2_PARAMS, POSEIDON2_BLS_3_PARAMS,
-            POSEIDON2_BLS_4_PARAMS, POSEIDON2_BLS_8_PARAMS,
-        },
-    };
-
-    type Scalar = FpBLS12;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let instances = vec![
-            Poseidon2::new(&POSEIDON2_BLS_2_PARAMS),
-            Poseidon2::new(&POSEIDON2_BLS_3_PARAMS),
-            Poseidon2::new(&POSEIDON2_BLS_4_PARAMS),
-            Poseidon2::new(&POSEIDON2_BLS_8_PARAMS),
-        ];
-        for instance in instances {
-            let t = instance.params.t;
-            for _ in 0..TESTRUNS {
-                let input1: Vec<Scalar> =
-                    (0..t).map(|_| random_scalar()).collect();
-
-                let mut input2: Vec<Scalar>;
-                loop {
-                    input2 = (0..t).map(|_| random_scalar()).collect();
-                    if input1 != input2 {
-                        break;
-                    }
-                }
-
-                let perm1 = instance.permutation(&input1);
-                let perm2 = instance.permutation(&input1);
-                let perm3 = instance.permutation(&input2);
-                assert_eq!(perm1, perm2);
-                assert_ne!(perm1, perm3);
-            }
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2_2 = Poseidon2::new(&POSEIDON2_BLS_2_PARAMS);
-        let mut input_2: Vec<Scalar> = vec![];
-        for i in 0..poseidon2_2.params.t {
-            input_2.push(Scalar::from(i as u64));
-        }
-        let perm_2 = poseidon2_2.permutation(&input_2);
-        assert_eq!(
-            perm_2[0],
-            fp_from_hex!(
-                "73c46dd530e248a87b61d19e67fa1b4ed30fc3d09f16531fe189fb945a15ce4e"
-            )
-        );
-        assert_eq!(
-            perm_2[1],
-            fp_from_hex!(
-                "1f0e305ee21c9366d5793b80251405032a3fee32b9dd0b5f4578262891b043b4"
-            )
-        );
-
-        let poseidon2_3 = Poseidon2::new(&POSEIDON2_BLS_3_PARAMS);
-        let mut input_3: Vec<Scalar> = vec![];
-        for i in 0..poseidon2_3.params.t {
-            input_3.push(Scalar::from(i as u64));
-        }
-        let perm_3 = poseidon2_3.permutation(&input_3);
-        assert_eq!(
-            perm_3[0],
-            fp_from_hex!(
-                "1b152349b1950b6a8ca75ee4407b6e26ca5cca5650534e56ef3fd45761fbf5f0"
-            )
-        );
-        assert_eq!(
-            perm_3[1],
-            fp_from_hex!(
-                "4c5793c87d51bdc2c08a32108437dc0000bd0275868f09ebc5f36919af5b3891"
-            )
-        );
-        assert_eq!(
-            perm_3[2],
-            fp_from_hex!(
-                "1fc8ed171e67902ca49863159fe5ba6325318843d13976143b8125f08b50dc6b"
-            )
-        );
-
-        let poseidon2_4 = Poseidon2::new(&POSEIDON2_BLS_4_PARAMS);
-        let mut input_4: Vec<Scalar> = vec![];
-        for i in 0..poseidon2_4.params.t {
-            input_4.push(Scalar::from(i as u64));
-        }
-        let perm_4 = poseidon2_4.permutation(&input_4);
-        assert_eq!(
-            perm_4[0],
-            fp_from_hex!(
-                "28ff6c4edf9768c08ae26290487e93449cc8bc155fc2fad92a344adceb3ada6d"
-            )
-        );
-        assert_eq!(
-            perm_4[1],
-            fp_from_hex!(
-                "0e56f2b6fad25075aa93560185b70e2b180ed7e269159c507c288b6747a0db2d"
-            )
-        );
-        assert_eq!(
-            perm_4[2],
-            fp_from_hex!(
-                "6d8196f28da6006bb89b3df94600acdc03d0ba7c2b0f3f4409a54c1db6bf30d0"
-            )
-        );
-        assert_eq!(
-            perm_4[3],
-            fp_from_hex!(
-                "07cfb49540ee456cce38b8a7d1a930a57ffc6660737f6589ef184c5e15334e36"
-            )
-        );
-    }
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
-mod poseidon2_tests_bn256 {
-    use super::*;
-    use crate::{
-        field::instance::FpBN256, fp_from_hex,
-        poseidon2::instance::bn256::POSEIDON2_BN256_PARAMS,
-    };
-
-    type Scalar = FpBN256;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_BN256_PARAMS);
-        let t = poseidon2.params.t;
-        for _ in 0..TESTRUNS {
-            let input1: Vec<Scalar> = (0..t).map(|_| random_scalar()).collect();
-
-            let mut input2: Vec<Scalar>;
-            loop {
-                input2 = (0..t).map(|_| random_scalar()).collect();
-                if input1 != input2 {
-                    break;
-                }
-            }
-
-            let perm1 = poseidon2.permutation(&input1);
-            let perm2 = poseidon2.permutation(&input1);
-            let perm3 = poseidon2.permutation(&input2);
-            assert_eq!(perm1, perm2);
-            assert_ne!(perm1, perm3);
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_BN256_PARAMS);
-        let mut input: Vec<Scalar> = vec![];
-        for i in 0..poseidon2.params.t {
-            input.push(Scalar::from(i as u64));
-        }
-        let perm = poseidon2.permutation(&input);
-        assert_eq!(
-            perm[0],
-            fp_from_hex!(
-                "0bb61d24daca55eebcb1929a82650f328134334da98ea4f847f760054f4a3033"
-            )
-        );
-        assert_eq!(
-            perm[1],
-            fp_from_hex!(
-                "303b6f7c86d043bfcbcc80214f26a30277a15d3f74ca654992defe7ff8d03570"
-            )
-        );
-        assert_eq!(
-            perm[2],
-            fp_from_hex!(
-                "1ed25194542b12eef8617361c3ba7c52e660b145994427cc86296242cf766ec8"
-            )
-        );
-    }
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
-mod poseidon2_tests_pallas {
-    use std::convert::TryFrom;
-
-    use super::*;
-    use crate::{
-        field::instance::FpPallas,
-        fp_from_hex,
-        poseidon2::instance::pallas::{
-            POSEIDON2_PALLAS_3_PARAMS, POSEIDON2_PALLAS_4_PARAMS,
-            POSEIDON2_PALLAS_8_PARAMS,
-        },
-    };
-
-    type Scalar = FpPallas;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let instances = vec![
-            Poseidon2::new(&POSEIDON2_PALLAS_3_PARAMS),
-            Poseidon2::new(&POSEIDON2_PALLAS_4_PARAMS),
-            Poseidon2::new(&POSEIDON2_PALLAS_8_PARAMS),
-        ];
-        for instance in instances {
-            let t = instance.params.t;
-            for _ in 0..TESTRUNS {
-                let input1: Vec<Scalar> =
-                    (0..t).map(|_| random_scalar()).collect();
-
-                let mut input2: Vec<Scalar>;
-                loop {
-                    input2 = (0..t).map(|_| random_scalar()).collect();
-                    if input1 != input2 {
-                        break;
-                    }
-                }
-
-                let perm1 = instance.permutation(&input1);
-                let perm2 = instance.permutation(&input1);
-                let perm3 = instance.permutation(&input2);
-                assert_eq!(perm1, perm2);
-                assert_ne!(perm1, perm3);
-            }
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_PALLAS_3_PARAMS);
-        let mut input: Vec<Scalar> = vec![];
-        for i in 0..poseidon2.params.t {
-            input.push(Scalar::from(i as u64));
-        }
-        let perm = poseidon2.permutation(&input);
-        assert_eq!(
-            perm[0],
-            fp_from_hex!(
-                "1a9b54c7512a914dd778282c44b3513fea7251420b9d95750baae059b2268d7a"
-            )
-        );
-        assert_eq!(
-            perm[1],
-            fp_from_hex!(
-                "1c48ea0994a7d7984ea338a54dbf0c8681f5af883fe988d59ba3380c9f7901fc"
-            )
-        );
-        assert_eq!(
-            perm[2],
-            fp_from_hex!(
-                "079ddd0a80a3e9414489b526a2770448964766685f4c4842c838f8a23120b401"
-            )
-        );
-    }
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
-mod poseidon2_tests_vesta {
-    use super::*;
-    use crate::{
-        field::instance::FpVesta, fp_from_hex,
-        poseidon2::instance::vesta::POSEIDON2_VESTA_PARAMS,
-    };
-
-    type Scalar = FpVesta;
-
-    static TESTRUNS: usize = 5;
-
-    #[test]
-    fn consistent_perm() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_VESTA_PARAMS);
-        let t = poseidon2.params.t;
-        for _ in 0..TESTRUNS {
-            let input1: Vec<Scalar> = (0..t).map(|_| random_scalar()).collect();
-
-            let mut input2: Vec<Scalar>;
-            loop {
-                input2 = (0..t).map(|_| random_scalar()).collect();
-                if input1 != input2 {
-                    break;
-                }
-            }
-
-            let perm1 = poseidon2.permutation(&input1);
-            let perm2 = poseidon2.permutation(&input1);
-            let perm3 = poseidon2.permutation(&input2);
-            assert_eq!(perm1, perm2);
-            assert_ne!(perm1, perm3);
-        }
-    }
-
-    #[test]
-    fn kats() {
-        let poseidon2 = Poseidon2::new(&POSEIDON2_VESTA_PARAMS);
-        let mut input: Vec<Scalar> = vec![];
-        for i in 0..poseidon2.params.t {
-            input.push(Scalar::from(i as u64));
-        }
-        let perm = poseidon2.permutation(&input);
-        assert_eq!(perm[0], fp_from_hex!("261ecbdfd62c617b82d297705f18c788fc9831b14a6a2b8f61229bef68ce2792"));
-        assert_eq!(perm[1], fp_from_hex!("2c76327e0b7653873263158cf8545c282364b183880fcdea93ca8526d518c66f"));
-        assert_eq!(perm[2], fp_from_hex!("262316c0ce5244838c75873299b59d763ae0849d2dd31bdc95caf7db1c2901bf"));
-    }
-
-    // TODO#q: test against `zkhash` implementation should be added
-}
-
-#[allow(unused_imports)]
-#[cfg(test)]
 pub fn random_scalar<F: PrimeField + crypto_bigint::Random>() -> F {
     let mut rng = rand::thread_rng();
     F::random(&mut rng)
